pyproject.toml
[project]
name = "htcli"
version = "0.1.0"
description = "CLI to interact with Hypertensor"
authors = [{ name = "ultrashiny", email = "vasyl.123.000@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9,<4"
dependencies = [
    "typer[all]>=0.15.3,<0.16.0",
    "click==8.0.4",
    "cryptography>=41.0.0",
    "multihash>=0.1.0",
    "protobuf==5.29.0",
    "six>=1.16.0",
    "mnemonic>=0.21,<0.22",
    "pathlib>=1.0.1,<2.0.0",
    "logging>=0.4.9.6,<0.5.0.0",
    "substrate-interface>=1.7.11,<2.0.0",
    "requests>=2.32.4",
    "websockets>=15.0.1",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.urls]
repository = "https://github.com/shiftlayer-llc/htcli"

[project.scripts]
htcli = "src.htcli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/htcli/**/*.py",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/htcli"]

# uv specific configuration
[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.17.0",
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]
